backend/dev-requirements.txt
-r requirements.txt
flake8>=3.7.8,<3.8.0
isort>=4.3.21,<4.4.0
pytest-django>=3.5.1,<3.6.0